Rename / relocate transport interface implementations
