+++ FreeRTOS-Plus/Demo/AWS/Ota_Windows_Simulator/Ota_Over_Mqtt_Demo/DemoTasks/OtaOverMqttDemoExample.c
@@ -72,7 +72,7 @@
 #include "backoff_algorithm.h"
 
 /* mbedTLS transport interface header.*/
-#include "using_mbedtls.h"
+#include "transport_mbedtls.h"
 
 /* OTA Library include. */
 #include "ota.h"
