+++ FreeRTOS-Plus/Demo/coreMQTT_Windows_Simulator/MQTT_Keep_Alive/DemoTasks/KeepAliveMQTTExample.c
@@ -68,7 +68,7 @@
 #include "backoff_algorithm.h"
 
 /* Transport interface include. */
-#include "using_plaintext.h"
+#include "transport_plaintext.h"
 
 /*-----------------------------------------------------------*/
