+++ FreeRTOS-Plus/Demo/coreMQTT_Windows_Simulator/MQTT_Mutual_Auth_wolfSSL/DemoTasks/MutualAuthMQTTExample.c
@@ -55,7 +55,7 @@
 #include "core_mqtt.h"
 
 /* Transport interface implementation include header for TLS. */
-#include "using_wolfSSL.h"
+#include "transport_wolfSSL.h"
 
 /*-----------------------------------------------------------*/
