+++ FreeRTOS-Plus/Source/Application-Protocols/network_transport/mbedtls_bio_freertos_plus_tcp.c
@@ -35,7 +35,7 @@
 
 /* mbed TLS includes. */
 #if !defined( MBEDTLS_CONFIG_FILE )
-    #include "config.h"
+    #include "mbedtls/mbedtls_config.h"
 #else
     #include MBEDTLS_CONFIG_FILE
 #endif
